+++ src/shared/components/index.ts
@@ -10,4 +10,4 @@ export * from './footer/footer.component';
 export * from './wishlist-button/wishlist-button.component';
 export * from './general-add-to-cart-icon/general-add-to-cart-icon.component';
 export * from './product-page-add-to-cart-icon/product-page-add-to-cart-icon.component';
-export * from './checkboxes';
+export * from './form-elements';
